% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_detection.R
\name{area_change}
\alias{area_change}
\title{change_detection}
\usage{
area_change(jul, apr)
}
\arguments{
\item{jul}{the first raster you want to use (newer date)}

\item{apr}{the second raster you want to use (older date)}
}
\description{
This function gives back a raster layer of the difference between two raster images (eg. surface changes during a year)
output: raster layer
}
\details{
this function does xxx
}
\keyword{change}
\keyword{detection,}
\keyword{layer,}
\keyword{raster}
\keyword{remote}
\keyword{sensing}
