% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryMap.R
\name{binaryMap}
\alias{binaryMap}
\title{binaryMap}
\usage{
binaryMap(raster, threshold)
}
\arguments{
\item{raster}{the raster you want to create a binary map out of (e.g index raste: NDVI, NDWI)}

\item{threshold}{defined threshold where each pixel gets assigned to one class}
}
\description{
This function calculates a binary map with the values 1 and 0 by setting a threshold.
output: raster layer
}
\details{
this function does xxx
}
\keyword{calculation,}
\keyword{index}
\keyword{remote}
\keyword{sensing}
\keyword{threshold,}
