% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_water_area.R
\name{calc_water_area}
\alias{calc_water_area}
\title{calc_water_area}
\usage{
calc_water_area(raster, res_x, res_y)
}
\arguments{
\item{raster}{the raster you want to use for the area calculation}

\item{res_x}{x resolution of the raster layer, default value is 30m}

\item{res_y}{y resolution of the raster layer, default value is 30m}
}
\description{
This function calculates the area in sqkm per pixel values of a binary raster imge (eg. water areas).
output: area in sqkm
}
\details{
this function does xxx
}
\keyword{calculation,}
\keyword{index}
\keyword{remote}
\keyword{sensing}
\keyword{threshold,}
