% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_band.R
\name{get_band}
\alias{get_band}
\title{get_band}
\usage{
get_band(band, image_folder, sensor_type)
}
\arguments{
\item{band}{the band you want to return}

\item{image_folder}{folder where your images are stored}

\item{sensor_type}{sentor type of your images}
}
\description{
This function returns back the NIR and GREEN bands for the satelites Landsat 4, 5 and 8.
output: raster layer
}
\details{
this function does xxx
}
\keyword{Landsat,}
\keyword{bands,}
\keyword{remote}
\keyword{sensing}
