% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_type.R
\name{get_sensor_type}
\alias{get_sensor_type}
\title{get_sensor_type}
\usage{
get_sensor_type(image_path)
}
\arguments{
\item{image_path}{uses the path where the image is located}
}
\description{
This function returns the satellite name based on the file name (only for Landsat 4, 5 and 8).
output: sensor type name
}
\details{
this function does xxx
}
\keyword{Landsat,}
\keyword{bands,}
\keyword{sensing}
\keyword{sensor,remote}
