% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDate.R
\name{getDate}
\alias{getDate}
\title{getDate}
\usage{
getDate(file)
}
\arguments{
\item{file}{path where the image is stored}
}
\description{
This function returns the recording date (only year and month, yyyymm) on an satellite raster image based on the original file na,e (only for Landsat 4, 5 and 8).
output: numeric value
}
\details{
this function does xxx
}
\keyword{date,}
\keyword{layer,}
\keyword{raster}
\keyword{remote}
\keyword{sensing}
