% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDWI.R
\name{NDWI}
\alias{NDWI}
\title{NDWI}
\usage{
NDWI(green, nir)
}
\arguments{
\item{green}{green band as input}

\item{nir}{nir band as input}
}
\description{
This function calculates the NDWI (Normaized Difference Vegetation Index) for Satellite image
output: raster layer
}
\details{
this function does xxx
}
\keyword{Landsat,}
\keyword{NDWI,}
\keyword{bands,}
\keyword{index,}
\keyword{remote}
\keyword{sensing}
\keyword{water}
